Started refactoring node code to follow these best practices: https://github.com/goldbergyoni/nodebestpractices#readme
